Make HotShot generic over ViewRunnerTYpe, remove ViewRunner
