+++ src/tasks/mod.rs
@@ -8,25 +8,16 @@ use async_compatibility_layer::{
 };
 use async_lock::RwLock;
 use async_trait::async_trait;
-use hotshot_consensus::{ConsensusApi, NextValidatingLeader, Replica, ValidatingLeader, ViewQueue};
-#[cfg(feature = "async-std-executor")]
-use hotshot_types::certificate::QuorumCertificate;
+use hotshot_consensus::ConsensusApi;
 use hotshot_types::{
     constants::LOOK_AHEAD,
-    data::{ValidatingLeaf, ValidatingProposal},
     message::MessageKind,
     traits::{
-        election::Election,
         network::NetworkingImplementation,
         node_implementation::{NodeImplementation, NodeType},
-        state::{
-            ConsensusType, SequencingConsensus, TestableBlock, TestableState, ValidatingConsensus,
-        },
     },
     ExecutionType,
 };
-#[allow(deprecated)]
-use nll::nll_todo::nll_todo;
 use std::{
     collections::HashMap,
     marker::PhantomData,
@@ -34,9 +25,9 @@ use std::{
         atomic::{AtomicBool, Ordering},
         Arc,
     },
-    time::{Duration, Instant},
+    time::Duration,
 };
-use tracing::{error, info, info_span, instrument, trace, Instrument};
+use tracing::{error, info, info_span, trace, Instrument};
 
 #[cfg(feature = "async-std-executor")]
 use async_std::task::{yield_now, JoinHandle};
@@ -149,10 +140,10 @@ struct TaskHandleInner {
 ///
 /// For a list of which tasks are being spawned, see this module's documentation.
 pub async fn spawn_all<TYPES: NodeType, I: NodeImplementation<TYPES>>(
-    hotshot: &HotShot<TYPES, I>,
+    hotshot: &HotShot<TYPES::ConsensusType, TYPES, I>,
 ) -> HotShotHandle<TYPES, I>
 where
-    ViewRunner<<TYPES as NodeType>::ConsensusType>: ViewRunnerType<TYPES, I>,
+    HotShot<TYPES::ConsensusType, TYPES, I>: ViewRunnerType<TYPES, I>,
 {
     let shut_down = Arc::new(AtomicBool::new(false));
     let started = Arc::new(AtomicBool::new(false));
@@ -218,224 +209,22 @@ where
     handle
 }
 
-// TODO (da) we may be able to get rid of `ViewRunner` and many where clauses by implementing
-// `ViewRunnerType` for `HotShot`.
-#[allow(clippy::missing_docs_in_private_items)]
-#[allow(missing_docs)]
-pub struct ViewRunner<CONSENSUS: ConsensusType> {
-    _pd: PhantomData<CONSENSUS>,
-}
-
 #[allow(clippy::missing_docs_in_private_items)]
 #[allow(missing_docs)]
 #[async_trait]
 pub trait ViewRunnerType<TYPES: NodeType, I: NodeImplementation<TYPES>> {
     /// Executes one view of consensus
-    async fn run_view(hotshot: HotShot<TYPES, I>) -> Result<(), ()>;
-}
-
-#[allow(clippy::too_many_lines)]
-#[async_trait]
-impl<
-        TYPES: NodeType<ConsensusType = ValidatingConsensus>,
-        ELECTION: Election<TYPES, LeafType = ValidatingLeaf<TYPES>>,
-        I: NodeImplementation<
-            TYPES,
-            Leaf = ValidatingLeaf<TYPES>,
-            Proposal = ValidatingProposal<TYPES, ELECTION>,
-        >,
-    > ViewRunnerType<TYPES, I> for ViewRunner<ValidatingConsensus>
-where
-    TYPES::StateType: TestableState,
-    TYPES::BlockType: TestableBlock,
-{
-    #[instrument(skip(hotshot), fields(id = hotshot.id), name = "Validating View Runner Task", level = "error")]
-    async fn run_view(hotshot: HotShot<TYPES, I>) -> Result<(), ()> {
-        let c_api = HotShotConsensusApi {
-            inner: hotshot.inner.clone(),
-        };
-        let start = Instant::now();
-        let metrics = Arc::clone(&hotshot.hotstuff.read().await.metrics);
-
-        // do book keeping on channel map
-        // TODO probably cleaner to separate this into a function
-        // e.g. insert the view and remove the last view
-        let mut send_to_replica = hotshot.replica_channel_map.write().await;
-        let replica_last_view: TYPES::Time = send_to_replica.cur_view;
-        // gc previous view's channel map
-        send_to_replica.channel_map.remove(&replica_last_view);
-        send_to_replica.cur_view += 1;
-        let replica_cur_view = send_to_replica.cur_view;
-        let ViewQueue {
-            sender_chan: send_replica,
-            receiver_chan: recv_replica,
-            has_received_proposal: _,
-        } = HotShot::<TYPES, I>::create_or_obtain_chan_from_write(
-            replica_cur_view,
-            send_to_replica,
-        )
-        .await;
-
-        let mut send_to_next_leader = hotshot.next_leader_channel_map.write().await;
-        let next_leader_last_view = send_to_next_leader.cur_view;
-        // gc previous view's channel map
-        send_to_next_leader
-            .channel_map
-            .remove(&next_leader_last_view);
-        send_to_next_leader.cur_view += 1;
-        let next_leader_cur_view = send_to_next_leader.cur_view;
-        let (send_next_leader, recv_next_leader) =
-            if c_api.is_leader(next_leader_cur_view + 1).await {
-                let vq = HotShot::<TYPES, I>::create_or_obtain_chan_from_write(
-                    next_leader_cur_view,
-                    send_to_next_leader,
-                )
-                .await;
-                (Some(vq.sender_chan), Some(vq.receiver_chan))
-            } else {
-                (None, None)
-            };
-
-        // increment consensus and start tasks
-
-        let (cur_view, high_qc, txns) = {
-            // OBTAIN write lock on consensus
-            let mut consensus = hotshot.hotstuff.write().await;
-            let cur_view = consensus.increment_view();
-            // make sure consistent
-            assert_eq!(cur_view, next_leader_cur_view);
-            assert_eq!(cur_view, replica_cur_view);
-            let high_qc = consensus.high_qc.clone();
-            let txns = consensus.transactions.clone();
-            // DROP write lock on consensus
-            drop(consensus);
-            (cur_view, high_qc, txns)
-        };
-
-        // notify networking to start worrying about the (`cur_view + LOOK_AHEAD`)th leader ahead of the current view
-        if hotshot
-            .send_network_lookup
-            .send(Some(cur_view))
-            .await
-            .is_err()
-        {
-            error!("Failed to initiate network lookup");
-        };
-
-        info!("Starting tasks for View {:?}!", cur_view);
-        metrics.current_view.set(*cur_view as usize);
-
-        let mut task_handles = Vec::new();
-
-        // replica always runs? TODO this will change once vrf integration is added
-        let replica = Replica {
-            id: hotshot.id,
-            consensus: hotshot.hotstuff.clone(),
-            proposal_collection_chan: recv_replica,
-            cur_view,
-            high_qc: high_qc.clone(),
-            api: c_api.clone(),
-        };
-        let replica_handle = async_spawn(async move { replica.run_view().await });
-        task_handles.push(replica_handle);
-
-        if c_api.is_leader(cur_view).await {
-            let leader = ValidatingLeader {
-                id: hotshot.id,
-                consensus: hotshot.hotstuff.clone(),
-                high_qc: high_qc.clone(),
-                cur_view,
-                transactions: txns,
-                api: c_api.clone(),
-                _pd: PhantomData,
-            };
-            let leader_handle = async_spawn(async move { leader.run_view().await });
-            task_handles.push(leader_handle);
-        }
-
-        if c_api.is_leader(cur_view + 1).await {
-            let next_leader = NextValidatingLeader {
-                id: hotshot.id,
-                generic_qc: high_qc,
-                // should be fine to unwrap here since the view numbers must be the same
-                vote_collection_chan: recv_next_leader.unwrap(),
-                cur_view,
-                api: c_api.clone(),
-                metrics,
-            };
-            let next_leader_handle = async_spawn(async move {
-                NextValidatingLeader::<HotShotConsensusApi<TYPES, I>, TYPES, _>::run_view(
-                    next_leader,
-                )
-                .await
-            });
-            task_handles.push(next_leader_handle);
-        }
-
-        let children_finished = futures::future::join_all(task_handles);
-
-        async_spawn({
-            let next_view_timeout = hotshot.inner.config.next_view_timeout;
-            let next_view_timeout = next_view_timeout;
-            let hotshot: HotShot<TYPES, I> = hotshot.clone();
-            async move {
-                async_sleep(Duration::from_millis(next_view_timeout)).await;
-                hotshot
-                    .timeout_view(cur_view, send_replica, send_next_leader)
-                    .await;
-            }
-        });
-
-        let results = children_finished.await;
-
-        // unwrap is fine since results must have >= 1 item(s)
-        #[cfg(feature = "async-std-executor")]
-        let high_qc = results
-            .into_iter()
-            .max_by_key(|qc: &QuorumCertificate<TYPES, I::Leaf>| qc.view_number)
-            .unwrap();
-        #[cfg(feature = "tokio-executor")]
-        let high_qc = results
-            .into_iter()
-            .filter_map(std::result::Result::ok)
-            .max_by_key(|qc| qc.view_number)
-            .unwrap();
-
-        #[cfg(not(any(feature = "async-std-executor", feature = "tokio-executor")))]
-        compile_error! {"Either feature \"async-std-executor\" or feature \"tokio-executor\" must be enabled for this crate."}
-
-        let mut consensus = hotshot.hotstuff.write().await;
-        consensus.high_qc = high_qc;
-        consensus
-            .metrics
-            .view_duration
-            .add_point(start.elapsed().as_secs_f64());
-        c_api.send_view_finished(consensus.cur_view).await;
-
-        info!("Returning from view {:?}!", cur_view);
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl<TYPES: NodeType<ConsensusType = SequencingConsensus>, I: NodeImplementation<TYPES>>
-    ViewRunnerType<TYPES, I> for ViewRunner<SequencingConsensus>
-{
-    // #[instrument]
-    async fn run_view(_hotshot: HotShot<TYPES, I>) -> Result<(), ()> {
-        #[allow(deprecated)]
-        nll_todo()
-    }
+    async fn run_view(hotshot: HotShot<TYPES::ConsensusType, TYPES, I>) -> Result<(), ()>;
 }
 
 /// main thread driving consensus
 pub async fn view_runner<TYPES: NodeType, I: NodeImplementation<TYPES>>(
-    hotshot: HotShot<TYPES, I>,
+    hotshot: HotShot<TYPES::ConsensusType, TYPES, I>,
     started: Arc<AtomicBool>,
     shut_down: Arc<AtomicBool>,
     run_once: Option<UnboundedReceiver<()>>,
 ) where
-    ViewRunner<<TYPES as NodeType>::ConsensusType>: ViewRunnerType<TYPES, I>,
+    HotShot<TYPES::ConsensusType, TYPES, I>: ViewRunnerType<TYPES, I>,
 {
     while !shut_down.load(Ordering::Relaxed) && !started.load(Ordering::Relaxed) {
         yield_now().await;
@@ -445,16 +234,13 @@ pub async fn view_runner<TYPES: NodeType, I: NodeImplementation<TYPES>>(
         if let Some(ref recv) = run_once {
             let _ = recv.recv().await;
         }
-        let _ = <ViewRunner<TYPES::ConsensusType> as ViewRunnerType<TYPES, I>>::run_view(
-            hotshot.clone(),
-        )
-        .await;
+        let _ = HotShot::<TYPES::ConsensusType, TYPES, I>::run_view(hotshot.clone()).await;
     }
 }
 
 /// Task to look up a node in the future as needed
 pub async fn network_lookup_task<TYPES: NodeType, I: NodeImplementation<TYPES>>(
-    hotshot: HotShot<TYPES, I>,
+    hotshot: HotShot<TYPES::ConsensusType, TYPES, I>,
     shut_down: Arc<AtomicBool>,
 ) {
     info!("Launching network lookup task");
@@ -516,11 +302,9 @@ pub async fn network_lookup_task<TYPES: NodeType, I: NodeImplementation<TYPES>>(
 
 /// Continually processes the incoming broadcast messages received on `hotshot.inner.networking`, redirecting them to `hotshot.handle_broadcast_*_message`.
 pub async fn network_broadcast_task<TYPES: NodeType, I: NodeImplementation<TYPES>>(
-    hotshot: HotShot<TYPES, I>,
+    hotshot: HotShot<TYPES::ConsensusType, TYPES, I>,
     shut_down: Arc<AtomicBool>,
-) where
-    ViewRunner<<TYPES as NodeType>::ConsensusType>: ViewRunnerType<TYPES, I>,
-{
+) {
     info!("Launching broadcast processing task");
     let networking = &hotshot.inner.networking;
     let mut incremental_backoff_ms = 10;
@@ -571,11 +355,9 @@ pub async fn network_broadcast_task<TYPES: NodeType, I: NodeImplementation<TYPES
 
 /// Continually processes the incoming direct messages received on `hotshot.inner.networking`, redirecting them to `hotshot.handle_direct_*_message`.
 pub async fn network_direct_task<TYPES: NodeType, I: NodeImplementation<TYPES>>(
-    hotshot: HotShot<TYPES, I>,
+    hotshot: HotShot<TYPES::ConsensusType, TYPES, I>,
     shut_down: Arc<AtomicBool>,
-) where
-    ViewRunner<<TYPES as NodeType>::ConsensusType>: ViewRunnerType<TYPES, I>,
-{
+) {
     info!("Launching direct processing task");
     let networking = &hotshot.inner.networking;
     let mut incremental_backoff_ms = 10;
@@ -618,11 +400,9 @@ pub async fn network_direct_task<TYPES: NodeType, I: NodeImplementation<TYPES>>(
 
 /// Runs a task that will call `hotshot.handle_network_change` whenever a change in the network is detected.
 pub async fn network_change_task<TYPES: NodeType, I: NodeImplementation<TYPES>>(
-    hotshot: HotShot<TYPES, I>,
+    hotshot: HotShot<TYPES::ConsensusType, TYPES, I>,
     shut_down: Arc<AtomicBool>,
-) where
-    ViewRunner<<TYPES as NodeType>::ConsensusType>: ViewRunnerType<TYPES, I>,
-{
+) {
     info!("Launching network change handler task");
     let networking = &hotshot.inner.networking;
     let mut incremental_backoff_ms = 10;
